decimal allow only in pos, migrations

+++ products/models.py
@@ -793,7 +793,7 @@ class PurchaseItem(models.Model):
         null=True,
         related_name='purchase_history'
     )
-    quantity = models.PositiveIntegerField()
+    quantity = models.DecimalField(max_digits=12, decimal_places=3)
     purchase_price = models.DecimalField(max_digits=10, decimal_places=2)  # Rate per unit
     total = models.DecimalField(max_digits=12, decimal_places=2)  # Qty * Rate
